+++ lib/src/rest/bloc/rest_event.dart
@@ -5,12 +5,28 @@ part 'rest_event.freezed.dart';
 @freezed
 abstract class RestEvent with _$RestEvent {
   factory RestEvent.get(String path,
-      {Map<String, String> params, Function fromJson}) = GetEvent;
-  factory RestEvent.post(String path, {String body, Function fromJson}) =
-      PostEvent;
-  factory RestEvent.put(String path, {String body, Function fromJson}) =
-      PutEvent;
-  factory RestEvent.patch(String path, {String body, Function fromJson}) =
-      PatchEvent;
-  factory RestEvent.delete(String path) = DeleteEvent;
+      {Function fromJson,
+      Map<String, String> headers,
+      Map<String, String> params}) = GetEvent;
+
+  factory RestEvent.post(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) = PostEvent;
+
+  factory RestEvent.put(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) = PutEvent;
+
+  factory RestEvent.patch(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) = PatchEvent;
+
+  factory RestEvent.delete(String path, {Map<String, String> headers}) =
+      DeleteEvent;
 }