+++ lib/src/rest/bloc/rest_event.freezed.dart
@@ -9,26 +9,38 @@ part of 'rest_event.dart';
 
 mixin _$RestEvent {
   String get path;
+  Map<String, String> get headers;
 
-  RestEvent copyWith({String path});
+  RestEvent copyWith({String path, Map<String, String> headers});
 
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required
-        Result get(String path, Map<String, String> params, Function fromJson),
-    @required Result post(String path, String body, Function fromJson),
-    @required Result put(String path, String body, Function fromJson),
-    @required Result patch(String path, String body, Function fromJson),
-    @required Result delete(String path),
+        Result get(String path, Function fromJson, Map<String, String> headers,
+            Map<String, String> params),
+    @required
+        Result post(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result put(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result patch(String path, Function fromJson,
+            Map<String, String> headers, String body, String contentType),
+    @required Result delete(String path, Map<String, String> headers),
   });
 
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
-    Result get(String path, Map<String, String> params, Function fromJson),
-    Result post(String path, String body, Function fromJson),
-    Result put(String path, String body, Function fromJson),
-    Result patch(String path, String body, Function fromJson),
-    Result delete(String path),
+    Result get(String path, Function fromJson, Map<String, String> headers,
+        Map<String, String> params),
+    Result post(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result put(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result patch(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result delete(String path, Map<String, String> headers),
     @required Result orElse(),
   });
 
@@ -55,41 +67,64 @@ mixin _$RestEvent {
 class _$RestEventTearOff {
   const _$RestEventTearOff();
 
-  GetEvent get(String path, {Map<String, String> params, Function fromJson}) {
+  GetEvent get(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      Map<String, String> params}) {
     return GetEvent(
       path,
-      params: params,
       fromJson: fromJson,
+      headers: headers,
+      params: params,
     );
   }
 
-  PostEvent post(String path, {String body, Function fromJson}) {
+  PostEvent post(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) {
     return PostEvent(
       path,
-      body: body,
       fromJson: fromJson,
+      headers: headers,
+      body: body,
+      contentType: contentType,
     );
   }
 
-  PutEvent put(String path, {String body, Function fromJson}) {
+  PutEvent put(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) {
     return PutEvent(
       path,
-      body: body,
       fromJson: fromJson,
+      headers: headers,
+      body: body,
+      contentType: contentType,
     );
   }
 
-  PatchEvent patch(String path, {String body, Function fromJson}) {
+  PatchEvent patch(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) {
     return PatchEvent(
       path,
-      body: body,
       fromJson: fromJson,
+      headers: headers,
+      body: body,
+      contentType: contentType,
     );
   }
 
-  DeleteEvent delete(String path) {
+  DeleteEvent delete(String path, {Map<String, String> headers}) {
     return DeleteEvent(
       path,
+      headers: headers,
     );
   }
 }
@@ -97,18 +132,21 @@ class _$RestEventTearOff {
 const $RestEvent = _$RestEventTearOff();
 
 class _$GetEvent implements GetEvent {
-  _$GetEvent(this.path, {this.params, this.fromJson}) : assert(path != null);
+  _$GetEvent(this.path, {this.fromJson, this.headers, this.params})
+      : assert(path != null);
 
   @override
   final String path;
   @override
-  final Map<String, String> params;
-  @override
   final Function fromJson;
+  @override
+  final Map<String, String> headers;
+  @override
+  final Map<String, String> params;
 
   @override
   String toString() {
-    return 'RestEvent.get(path: $path, params: $params, fromJson: $fromJson)';
+    return 'RestEvent.get(path: $path, fromJson: $fromJson, headers: $headers, params: $params)';
   }
 
   @override
@@ -117,30 +155,37 @@ class _$GetEvent implements GetEvent {
         (other is GetEvent &&
             (identical(other.path, path) ||
                 const DeepCollectionEquality().equals(other.path, path)) &&
-            (identical(other.params, params) ||
-                const DeepCollectionEquality().equals(other.params, params)) &&
             (identical(other.fromJson, fromJson) ||
                 const DeepCollectionEquality()
-                    .equals(other.fromJson, fromJson)));
+                    .equals(other.fromJson, fromJson)) &&
+            (identical(other.headers, headers) ||
+                const DeepCollectionEquality()
+                    .equals(other.headers, headers)) &&
+            (identical(other.params, params) ||
+                const DeepCollectionEquality().equals(other.params, params)));
   }
 
   @override
   int get hashCode =>
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(path) ^
-      const DeepCollectionEquality().hash(params) ^
-      const DeepCollectionEquality().hash(fromJson);
+      const DeepCollectionEquality().hash(fromJson) ^
+      const DeepCollectionEquality().hash(headers) ^
+      const DeepCollectionEquality().hash(params);
 
   @override
   _$GetEvent copyWith({
     Object path = freezed,
-    Object params = freezed,
     Object fromJson = freezed,
+    Object headers = freezed,
+    Object params = freezed,
   }) {
     return _$GetEvent(
       path == freezed ? this.path : path as String,
-      params: params == freezed ? this.params : params as Map<String, String>,
       fromJson: fromJson == freezed ? this.fromJson : fromJson as Function,
+      headers:
+          headers == freezed ? this.headers : headers as Map<String, String>,
+      params: params == freezed ? this.params : params as Map<String, String>,
     );
   }
 
@@ -148,33 +193,44 @@ class _$GetEvent implements GetEvent {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required
-        Result get(String path, Map<String, String> params, Function fromJson),
-    @required Result post(String path, String body, Function fromJson),
-    @required Result put(String path, String body, Function fromJson),
-    @required Result patch(String path, String body, Function fromJson),
-    @required Result delete(String path),
+        Result get(String path, Function fromJson, Map<String, String> headers,
+            Map<String, String> params),
+    @required
+        Result post(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result put(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result patch(String path, Function fromJson,
+            Map<String, String> headers, String body, String contentType),
+    @required Result delete(String path, Map<String, String> headers),
   }) {
     assert(get != null);
     assert(post != null);
     assert(put != null);
     assert(patch != null);
     assert(delete != null);
-    return get(path, params, fromJson);
+    return get(path, fromJson, headers, params);
   }
 
   @override
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
-    Result get(String path, Map<String, String> params, Function fromJson),
-    Result post(String path, String body, Function fromJson),
-    Result put(String path, String body, Function fromJson),
-    Result patch(String path, String body, Function fromJson),
-    Result delete(String path),
+    Result get(String path, Function fromJson, Map<String, String> headers,
+        Map<String, String> params),
+    Result post(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result put(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result patch(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result delete(String path, Map<String, String> headers),
     @required Result orElse(),
   }) {
     assert(orElse != null);
     if (get != null) {
-      return get(path, params, fromJson);
+      return get(path, fromJson, headers, params);
     }
     return orElse();
   }
@@ -216,31 +272,44 @@ class _$GetEvent implements GetEvent {
 
 abstract class GetEvent implements RestEvent {
   factory GetEvent(String path,
-      {Map<String, String> params, Function fromJson}) = _$GetEvent;
+      {Function fromJson,
+      Map<String, String> headers,
+      Map<String, String> params}) = _$GetEvent;
 
   @override
   String get path;
-  Map<String, String> get params;
   Function get fromJson;
+  @override
+  Map<String, String> get headers;
+  Map<String, String> get params;
 
   @override
   GetEvent copyWith(
-      {String path, Map<String, String> params, Function fromJson});
+      {String path,
+      Function fromJson,
+      Map<String, String> headers,
+      Map<String, String> params});
 }
 
 class _$PostEvent implements PostEvent {
-  _$PostEvent(this.path, {this.body, this.fromJson}) : assert(path != null);
+  _$PostEvent(this.path,
+      {this.fromJson, this.headers, this.body, this.contentType})
+      : assert(path != null);
 
   @override
   final String path;
   @override
+  final Function fromJson;
+  @override
+  final Map<String, String> headers;
+  @override
   final String body;
   @override
-  final Function fromJson;
+  final String contentType;
 
   @override
   String toString() {
-    return 'RestEvent.post(path: $path, body: $body, fromJson: $fromJson)';
+    return 'RestEvent.post(path: $path, fromJson: $fromJson, headers: $headers, body: $body, contentType: $contentType)';
   }
 
   @override
@@ -249,30 +318,44 @@ class _$PostEvent implements PostEvent {
         (other is PostEvent &&
             (identical(other.path, path) ||
                 const DeepCollectionEquality().equals(other.path, path)) &&
-            (identical(other.body, body) ||
-                const DeepCollectionEquality().equals(other.body, body)) &&
             (identical(other.fromJson, fromJson) ||
                 const DeepCollectionEquality()
-                    .equals(other.fromJson, fromJson)));
+                    .equals(other.fromJson, fromJson)) &&
+            (identical(other.headers, headers) ||
+                const DeepCollectionEquality()
+                    .equals(other.headers, headers)) &&
+            (identical(other.body, body) ||
+                const DeepCollectionEquality().equals(other.body, body)) &&
+            (identical(other.contentType, contentType) ||
+                const DeepCollectionEquality()
+                    .equals(other.contentType, contentType)));
   }
 
   @override
   int get hashCode =>
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(path) ^
+      const DeepCollectionEquality().hash(fromJson) ^
+      const DeepCollectionEquality().hash(headers) ^
       const DeepCollectionEquality().hash(body) ^
-      const DeepCollectionEquality().hash(fromJson);
+      const DeepCollectionEquality().hash(contentType);
 
   @override
   _$PostEvent copyWith({
     Object path = freezed,
-    Object body = freezed,
     Object fromJson = freezed,
+    Object headers = freezed,
+    Object body = freezed,
+    Object contentType = freezed,
   }) {
     return _$PostEvent(
       path == freezed ? this.path : path as String,
-      body: body == freezed ? this.body : body as String,
       fromJson: fromJson == freezed ? this.fromJson : fromJson as Function,
+      headers:
+          headers == freezed ? this.headers : headers as Map<String, String>,
+      body: body == freezed ? this.body : body as String,
+      contentType:
+          contentType == freezed ? this.contentType : contentType as String,
     );
   }
 
@@ -280,33 +363,44 @@ class _$PostEvent implements PostEvent {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required
-        Result get(String path, Map<String, String> params, Function fromJson),
-    @required Result post(String path, String body, Function fromJson),
-    @required Result put(String path, String body, Function fromJson),
-    @required Result patch(String path, String body, Function fromJson),
-    @required Result delete(String path),
+        Result get(String path, Function fromJson, Map<String, String> headers,
+            Map<String, String> params),
+    @required
+        Result post(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result put(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result patch(String path, Function fromJson,
+            Map<String, String> headers, String body, String contentType),
+    @required Result delete(String path, Map<String, String> headers),
   }) {
     assert(get != null);
     assert(post != null);
     assert(put != null);
     assert(patch != null);
     assert(delete != null);
-    return post(path, body, fromJson);
+    return post(path, fromJson, headers, body, contentType);
   }
 
   @override
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
-    Result get(String path, Map<String, String> params, Function fromJson),
-    Result post(String path, String body, Function fromJson),
-    Result put(String path, String body, Function fromJson),
-    Result patch(String path, String body, Function fromJson),
-    Result delete(String path),
+    Result get(String path, Function fromJson, Map<String, String> headers,
+        Map<String, String> params),
+    Result post(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result put(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result patch(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result delete(String path, Map<String, String> headers),
     @required Result orElse(),
   }) {
     assert(orElse != null);
     if (post != null) {
-      return post(path, body, fromJson);
+      return post(path, fromJson, headers, body, contentType);
     }
     return orElse();
   }
@@ -347,31 +441,48 @@ class _$PostEvent implements PostEvent {
 }
 
 abstract class PostEvent implements RestEvent {
-  factory PostEvent(String path, {String body, Function fromJson}) =
-      _$PostEvent;
+  factory PostEvent(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) = _$PostEvent;
 
   @override
   String get path;
-  String get body;
   Function get fromJson;
+  @override
+  Map<String, String> get headers;
+  String get body;
+  String get contentType;
 
   @override
-  PostEvent copyWith({String path, String body, Function fromJson});
+  PostEvent copyWith(
+      {String path,
+      Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType});
 }
 
 class _$PutEvent implements PutEvent {
-  _$PutEvent(this.path, {this.body, this.fromJson}) : assert(path != null);
+  _$PutEvent(this.path,
+      {this.fromJson, this.headers, this.body, this.contentType})
+      : assert(path != null);
 
   @override
   final String path;
   @override
+  final Function fromJson;
+  @override
+  final Map<String, String> headers;
+  @override
   final String body;
   @override
-  final Function fromJson;
+  final String contentType;
 
   @override
   String toString() {
-    return 'RestEvent.put(path: $path, body: $body, fromJson: $fromJson)';
+    return 'RestEvent.put(path: $path, fromJson: $fromJson, headers: $headers, body: $body, contentType: $contentType)';
   }
 
   @override
@@ -380,30 +491,44 @@ class _$PutEvent implements PutEvent {
         (other is PutEvent &&
             (identical(other.path, path) ||
                 const DeepCollectionEquality().equals(other.path, path)) &&
-            (identical(other.body, body) ||
-                const DeepCollectionEquality().equals(other.body, body)) &&
             (identical(other.fromJson, fromJson) ||
                 const DeepCollectionEquality()
-                    .equals(other.fromJson, fromJson)));
+                    .equals(other.fromJson, fromJson)) &&
+            (identical(other.headers, headers) ||
+                const DeepCollectionEquality()
+                    .equals(other.headers, headers)) &&
+            (identical(other.body, body) ||
+                const DeepCollectionEquality().equals(other.body, body)) &&
+            (identical(other.contentType, contentType) ||
+                const DeepCollectionEquality()
+                    .equals(other.contentType, contentType)));
   }
 
   @override
   int get hashCode =>
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(path) ^
+      const DeepCollectionEquality().hash(fromJson) ^
+      const DeepCollectionEquality().hash(headers) ^
       const DeepCollectionEquality().hash(body) ^
-      const DeepCollectionEquality().hash(fromJson);
+      const DeepCollectionEquality().hash(contentType);
 
   @override
   _$PutEvent copyWith({
     Object path = freezed,
-    Object body = freezed,
     Object fromJson = freezed,
+    Object headers = freezed,
+    Object body = freezed,
+    Object contentType = freezed,
   }) {
     return _$PutEvent(
       path == freezed ? this.path : path as String,
-      body: body == freezed ? this.body : body as String,
       fromJson: fromJson == freezed ? this.fromJson : fromJson as Function,
+      headers:
+          headers == freezed ? this.headers : headers as Map<String, String>,
+      body: body == freezed ? this.body : body as String,
+      contentType:
+          contentType == freezed ? this.contentType : contentType as String,
     );
   }
 
@@ -411,33 +536,44 @@ class _$PutEvent implements PutEvent {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required
-        Result get(String path, Map<String, String> params, Function fromJson),
-    @required Result post(String path, String body, Function fromJson),
-    @required Result put(String path, String body, Function fromJson),
-    @required Result patch(String path, String body, Function fromJson),
-    @required Result delete(String path),
+        Result get(String path, Function fromJson, Map<String, String> headers,
+            Map<String, String> params),
+    @required
+        Result post(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result put(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result patch(String path, Function fromJson,
+            Map<String, String> headers, String body, String contentType),
+    @required Result delete(String path, Map<String, String> headers),
   }) {
     assert(get != null);
     assert(post != null);
     assert(put != null);
     assert(patch != null);
     assert(delete != null);
-    return put(path, body, fromJson);
+    return put(path, fromJson, headers, body, contentType);
   }
 
   @override
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
-    Result get(String path, Map<String, String> params, Function fromJson),
-    Result post(String path, String body, Function fromJson),
-    Result put(String path, String body, Function fromJson),
-    Result patch(String path, String body, Function fromJson),
-    Result delete(String path),
+    Result get(String path, Function fromJson, Map<String, String> headers,
+        Map<String, String> params),
+    Result post(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result put(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result patch(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result delete(String path, Map<String, String> headers),
     @required Result orElse(),
   }) {
     assert(orElse != null);
     if (put != null) {
-      return put(path, body, fromJson);
+      return put(path, fromJson, headers, body, contentType);
     }
     return orElse();
   }
@@ -478,30 +614,48 @@ class _$PutEvent implements PutEvent {
 }
 
 abstract class PutEvent implements RestEvent {
-  factory PutEvent(String path, {String body, Function fromJson}) = _$PutEvent;
+  factory PutEvent(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) = _$PutEvent;
 
   @override
   String get path;
-  String get body;
   Function get fromJson;
+  @override
+  Map<String, String> get headers;
+  String get body;
+  String get contentType;
 
   @override
-  PutEvent copyWith({String path, String body, Function fromJson});
+  PutEvent copyWith(
+      {String path,
+      Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType});
 }
 
 class _$PatchEvent implements PatchEvent {
-  _$PatchEvent(this.path, {this.body, this.fromJson}) : assert(path != null);
+  _$PatchEvent(this.path,
+      {this.fromJson, this.headers, this.body, this.contentType})
+      : assert(path != null);
 
   @override
   final String path;
   @override
+  final Function fromJson;
+  @override
+  final Map<String, String> headers;
+  @override
   final String body;
   @override
-  final Function fromJson;
+  final String contentType;
 
   @override
   String toString() {
-    return 'RestEvent.patch(path: $path, body: $body, fromJson: $fromJson)';
+    return 'RestEvent.patch(path: $path, fromJson: $fromJson, headers: $headers, body: $body, contentType: $contentType)';
   }
 
   @override
@@ -510,30 +664,44 @@ class _$PatchEvent implements PatchEvent {
         (other is PatchEvent &&
             (identical(other.path, path) ||
                 const DeepCollectionEquality().equals(other.path, path)) &&
-            (identical(other.body, body) ||
-                const DeepCollectionEquality().equals(other.body, body)) &&
             (identical(other.fromJson, fromJson) ||
                 const DeepCollectionEquality()
-                    .equals(other.fromJson, fromJson)));
+                    .equals(other.fromJson, fromJson)) &&
+            (identical(other.headers, headers) ||
+                const DeepCollectionEquality()
+                    .equals(other.headers, headers)) &&
+            (identical(other.body, body) ||
+                const DeepCollectionEquality().equals(other.body, body)) &&
+            (identical(other.contentType, contentType) ||
+                const DeepCollectionEquality()
+                    .equals(other.contentType, contentType)));
   }
 
   @override
   int get hashCode =>
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(path) ^
+      const DeepCollectionEquality().hash(fromJson) ^
+      const DeepCollectionEquality().hash(headers) ^
       const DeepCollectionEquality().hash(body) ^
-      const DeepCollectionEquality().hash(fromJson);
+      const DeepCollectionEquality().hash(contentType);
 
   @override
   _$PatchEvent copyWith({
     Object path = freezed,
-    Object body = freezed,
     Object fromJson = freezed,
+    Object headers = freezed,
+    Object body = freezed,
+    Object contentType = freezed,
   }) {
     return _$PatchEvent(
       path == freezed ? this.path : path as String,
-      body: body == freezed ? this.body : body as String,
       fromJson: fromJson == freezed ? this.fromJson : fromJson as Function,
+      headers:
+          headers == freezed ? this.headers : headers as Map<String, String>,
+      body: body == freezed ? this.body : body as String,
+      contentType:
+          contentType == freezed ? this.contentType : contentType as String,
     );
   }
 
@@ -541,33 +709,44 @@ class _$PatchEvent implements PatchEvent {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required
-        Result get(String path, Map<String, String> params, Function fromJson),
-    @required Result post(String path, String body, Function fromJson),
-    @required Result put(String path, String body, Function fromJson),
-    @required Result patch(String path, String body, Function fromJson),
-    @required Result delete(String path),
+        Result get(String path, Function fromJson, Map<String, String> headers,
+            Map<String, String> params),
+    @required
+        Result post(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result put(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result patch(String path, Function fromJson,
+            Map<String, String> headers, String body, String contentType),
+    @required Result delete(String path, Map<String, String> headers),
   }) {
     assert(get != null);
     assert(post != null);
     assert(put != null);
     assert(patch != null);
     assert(delete != null);
-    return patch(path, body, fromJson);
+    return patch(path, fromJson, headers, body, contentType);
   }
 
   @override
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
-    Result get(String path, Map<String, String> params, Function fromJson),
-    Result post(String path, String body, Function fromJson),
-    Result put(String path, String body, Function fromJson),
-    Result patch(String path, String body, Function fromJson),
-    Result delete(String path),
+    Result get(String path, Function fromJson, Map<String, String> headers,
+        Map<String, String> params),
+    Result post(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result put(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result patch(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result delete(String path, Map<String, String> headers),
     @required Result orElse(),
   }) {
     assert(orElse != null);
     if (patch != null) {
-      return patch(path, body, fromJson);
+      return patch(path, fromJson, headers, body, contentType);
     }
     return orElse();
   }
@@ -608,27 +787,40 @@ class _$PatchEvent implements PatchEvent {
 }
 
 abstract class PatchEvent implements RestEvent {
-  factory PatchEvent(String path, {String body, Function fromJson}) =
-      _$PatchEvent;
+  factory PatchEvent(String path,
+      {Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType}) = _$PatchEvent;
 
   @override
   String get path;
-  String get body;
   Function get fromJson;
+  @override
+  Map<String, String> get headers;
+  String get body;
+  String get contentType;
 
   @override
-  PatchEvent copyWith({String path, String body, Function fromJson});
+  PatchEvent copyWith(
+      {String path,
+      Function fromJson,
+      Map<String, String> headers,
+      String body,
+      String contentType});
 }
 
 class _$DeleteEvent implements DeleteEvent {
-  _$DeleteEvent(this.path) : assert(path != null);
+  _$DeleteEvent(this.path, {this.headers}) : assert(path != null);
 
   @override
   final String path;
+  @override
+  final Map<String, String> headers;
 
   @override
   String toString() {
-    return 'RestEvent.delete(path: $path)';
+    return 'RestEvent.delete(path: $path, headers: $headers)';
   }
 
   @override
@@ -636,19 +828,26 @@ class _$DeleteEvent implements DeleteEvent {
     return identical(this, other) ||
         (other is DeleteEvent &&
             (identical(other.path, path) ||
-                const DeepCollectionEquality().equals(other.path, path)));
+                const DeepCollectionEquality().equals(other.path, path)) &&
+            (identical(other.headers, headers) ||
+                const DeepCollectionEquality().equals(other.headers, headers)));
   }
 
   @override
   int get hashCode =>
-      runtimeType.hashCode ^ const DeepCollectionEquality().hash(path);
+      runtimeType.hashCode ^
+      const DeepCollectionEquality().hash(path) ^
+      const DeepCollectionEquality().hash(headers);
 
   @override
   _$DeleteEvent copyWith({
     Object path = freezed,
+    Object headers = freezed,
   }) {
     return _$DeleteEvent(
       path == freezed ? this.path : path as String,
+      headers:
+          headers == freezed ? this.headers : headers as Map<String, String>,
     );
   }
 
@@ -656,33 +855,44 @@ class _$DeleteEvent implements DeleteEvent {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required
-        Result get(String path, Map<String, String> params, Function fromJson),
-    @required Result post(String path, String body, Function fromJson),
-    @required Result put(String path, String body, Function fromJson),
-    @required Result patch(String path, String body, Function fromJson),
-    @required Result delete(String path),
+        Result get(String path, Function fromJson, Map<String, String> headers,
+            Map<String, String> params),
+    @required
+        Result post(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result put(String path, Function fromJson, Map<String, String> headers,
+            String body, String contentType),
+    @required
+        Result patch(String path, Function fromJson,
+            Map<String, String> headers, String body, String contentType),
+    @required Result delete(String path, Map<String, String> headers),
   }) {
     assert(get != null);
     assert(post != null);
     assert(put != null);
     assert(patch != null);
     assert(delete != null);
-    return delete(path);
+    return delete(path, headers);
   }
 
   @override
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
-    Result get(String path, Map<String, String> params, Function fromJson),
-    Result post(String path, String body, Function fromJson),
-    Result put(String path, String body, Function fromJson),
-    Result patch(String path, String body, Function fromJson),
-    Result delete(String path),
+    Result get(String path, Function fromJson, Map<String, String> headers,
+        Map<String, String> params),
+    Result post(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result put(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result patch(String path, Function fromJson, Map<String, String> headers,
+        String body, String contentType),
+    Result delete(String path, Map<String, String> headers),
     @required Result orElse(),
   }) {
     assert(orElse != null);
     if (delete != null) {
-      return delete(path);
+      return delete(path, headers);
     }
     return orElse();
   }
@@ -723,11 +933,14 @@ class _$DeleteEvent implements DeleteEvent {
 }
 
 abstract class DeleteEvent implements RestEvent {
-  factory DeleteEvent(String path) = _$DeleteEvent;
+  factory DeleteEvent(String path, {Map<String, String> headers}) =
+      _$DeleteEvent;
 
   @override
   String get path;
+  @override
+  Map<String, String> get headers;
 
   @override
-  DeleteEvent copyWith({String path});
+  DeleteEvent copyWith({String path, Map<String, String> headers});
 }