+++ lib/src/rest/bloc/rest_state.freezed.dart
@@ -11,17 +11,20 @@ mixin _$RestState {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required Result uninitialized(),
-    @required Result error(String message),
+    @required Result error(String message, String humanMessage),
     @required Result loading(),
-    @required Result loaded(dynamic data, String lastPath, String timestamp),
+    @required
+        Result loaded(Map<String, String> headers, String lastPath,
+            String timestamp, dynamic data),
   });
 
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
     Result uninitialized(),
-    Result error(String message),
+    Result error(String message, String humanMessage),
     Result loading(),
-    Result loaded(dynamic data, String lastPath, String timestamp),
+    Result loaded(Map<String, String> headers, String lastPath,
+        String timestamp, dynamic data),
     @required Result orElse(),
   });
 
@@ -50,9 +53,10 @@ class _$RestStateTearOff {
     return UninitializedRestState();
   }
 
-  ErrorRestState error({String message}) {
+  ErrorRestState error({String message, String humanMessage}) {
     return ErrorRestState(
       message: message,
+      humanMessage: humanMessage,
     );
   }
 
@@ -60,11 +64,16 @@ class _$RestStateTearOff {
     return LoadingRestState();
   }
 
-  LoadedRestState loaded({dynamic data, String lastPath, String timestamp}) {
+  LoadedRestState loaded(
+      {Map<String, String> headers,
+      String lastPath,
+      String timestamp,
+      dynamic data}) {
     return LoadedRestState(
-      data: data,
+      headers: headers,
       lastPath: lastPath,
       timestamp: timestamp,
+      data: data,
     );
   }
 }
@@ -91,9 +100,11 @@ class _$UninitializedRestState implements UninitializedRestState {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required Result uninitialized(),
-    @required Result error(String message),
+    @required Result error(String message, String humanMessage),
     @required Result loading(),
-    @required Result loaded(dynamic data, String lastPath, String timestamp),
+    @required
+        Result loaded(Map<String, String> headers, String lastPath,
+            String timestamp, dynamic data),
   }) {
     assert(uninitialized != null);
     assert(error != null);
@@ -106,9 +117,10 @@ class _$UninitializedRestState implements UninitializedRestState {
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
     Result uninitialized(),
-    Result error(String message),
+    Result error(String message, String humanMessage),
     Result loading(),
-    Result loaded(dynamic data, String lastPath, String timestamp),
+    Result loaded(Map<String, String> headers, String lastPath,
+        String timestamp, dynamic data),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -155,14 +167,16 @@ abstract class UninitializedRestState implements RestState {
 }
 
 class _$ErrorRestState implements ErrorRestState {
-  _$ErrorRestState({this.message});
+  _$ErrorRestState({this.message, this.humanMessage});
 
   @override
   final String message;
+  @override
+  final String humanMessage;
 
   @override
   String toString() {
-    return 'RestState.error(message: $message)';
+    return 'RestState.error(message: $message, humanMessage: $humanMessage)';
   }
 
   @override
@@ -170,19 +184,28 @@ class _$ErrorRestState implements ErrorRestState {
     return identical(this, other) ||
         (other is ErrorRestState &&
             (identical(other.message, message) ||
-                const DeepCollectionEquality().equals(other.message, message)));
+                const DeepCollectionEquality()
+                    .equals(other.message, message)) &&
+            (identical(other.humanMessage, humanMessage) ||
+                const DeepCollectionEquality()
+                    .equals(other.humanMessage, humanMessage)));
   }
 
   @override
   int get hashCode =>
-      runtimeType.hashCode ^ const DeepCollectionEquality().hash(message);
+      runtimeType.hashCode ^
+      const DeepCollectionEquality().hash(message) ^
+      const DeepCollectionEquality().hash(humanMessage);
 
   @override
   _$ErrorRestState copyWith({
     Object message = freezed,
+    Object humanMessage = freezed,
   }) {
     return _$ErrorRestState(
       message: message == freezed ? this.message : message as String,
+      humanMessage:
+          humanMessage == freezed ? this.humanMessage : humanMessage as String,
     );
   }
 
@@ -190,29 +213,32 @@ class _$ErrorRestState implements ErrorRestState {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required Result uninitialized(),
-    @required Result error(String message),
+    @required Result error(String message, String humanMessage),
     @required Result loading(),
-    @required Result loaded(dynamic data, String lastPath, String timestamp),
+    @required
+        Result loaded(Map<String, String> headers, String lastPath,
+            String timestamp, dynamic data),
   }) {
     assert(uninitialized != null);
     assert(error != null);
     assert(loading != null);
     assert(loaded != null);
-    return error(message);
+    return error(message, humanMessage);
   }
 
   @override
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
     Result uninitialized(),
-    Result error(String message),
+    Result error(String message, String humanMessage),
     Result loading(),
-    Result loaded(dynamic data, String lastPath, String timestamp),
+    Result loaded(Map<String, String> headers, String lastPath,
+        String timestamp, dynamic data),
     @required Result orElse(),
   }) {
     assert(orElse != null);
     if (error != null) {
-      return error(message);
+      return error(message, humanMessage);
     }
     return orElse();
   }
@@ -250,11 +276,13 @@ class _$ErrorRestState implements ErrorRestState {
 }
 
 abstract class ErrorRestState implements RestState {
-  factory ErrorRestState({String message}) = _$ErrorRestState;
+  factory ErrorRestState({String message, String humanMessage}) =
+      _$ErrorRestState;
 
   String get message;
+  String get humanMessage;
 
-  ErrorRestState copyWith({String message});
+  ErrorRestState copyWith({String message, String humanMessage});
 }
 
 class _$LoadingRestState implements LoadingRestState {
@@ -277,9 +305,11 @@ class _$LoadingRestState implements LoadingRestState {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required Result uninitialized(),
-    @required Result error(String message),
+    @required Result error(String message, String humanMessage),
     @required Result loading(),
-    @required Result loaded(dynamic data, String lastPath, String timestamp),
+    @required
+        Result loaded(Map<String, String> headers, String lastPath,
+            String timestamp, dynamic data),
   }) {
     assert(uninitialized != null);
     assert(error != null);
@@ -292,9 +322,10 @@ class _$LoadingRestState implements LoadingRestState {
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
     Result uninitialized(),
-    Result error(String message),
+    Result error(String message, String humanMessage),
     Result loading(),
-    Result loaded(dynamic data, String lastPath, String timestamp),
+    Result loaded(Map<String, String> headers, String lastPath,
+        String timestamp, dynamic data),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -341,51 +372,60 @@ abstract class LoadingRestState implements RestState {
 }
 
 class _$LoadedRestState implements LoadedRestState {
-  _$LoadedRestState({this.data, this.lastPath, this.timestamp});
+  _$LoadedRestState({this.headers, this.lastPath, this.timestamp, this.data});
 
   @override
-  final dynamic data;
+  final Map<String, String> headers;
   @override
   final String lastPath;
   @override
   final String timestamp;
+  @override
+  final dynamic data;
 
   @override
   String toString() {
-    return 'RestState.loaded(data: $data, lastPath: $lastPath, timestamp: $timestamp)';
+    return 'RestState.loaded(headers: $headers, lastPath: $lastPath, timestamp: $timestamp, data: $data)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other is LoadedRestState &&
-            (identical(other.data, data) ||
-                const DeepCollectionEquality().equals(other.data, data)) &&
+            (identical(other.headers, headers) ||
+                const DeepCollectionEquality()
+                    .equals(other.headers, headers)) &&
             (identical(other.lastPath, lastPath) ||
                 const DeepCollectionEquality()
                     .equals(other.lastPath, lastPath)) &&
             (identical(other.timestamp, timestamp) ||
                 const DeepCollectionEquality()
-                    .equals(other.timestamp, timestamp)));
+                    .equals(other.timestamp, timestamp)) &&
+            (identical(other.data, data) ||
+                const DeepCollectionEquality().equals(other.data, data)));
   }
 
   @override
   int get hashCode =>
       runtimeType.hashCode ^
-      const DeepCollectionEquality().hash(data) ^
+      const DeepCollectionEquality().hash(headers) ^
       const DeepCollectionEquality().hash(lastPath) ^
-      const DeepCollectionEquality().hash(timestamp);
+      const DeepCollectionEquality().hash(timestamp) ^
+      const DeepCollectionEquality().hash(data);
 
   @override
   _$LoadedRestState copyWith({
-    Object data = freezed,
+    Object headers = freezed,
     Object lastPath = freezed,
     Object timestamp = freezed,
+    Object data = freezed,
   }) {
     return _$LoadedRestState(
-      data: data == freezed ? this.data : data as dynamic,
+      headers:
+          headers == freezed ? this.headers : headers as Map<String, String>,
       lastPath: lastPath == freezed ? this.lastPath : lastPath as String,
       timestamp: timestamp == freezed ? this.timestamp : timestamp as String,
+      data: data == freezed ? this.data : data as dynamic,
     );
   }
 
@@ -393,29 +433,32 @@ class _$LoadedRestState implements LoadedRestState {
   @optionalTypeArgs
   Result when<Result extends Object>({
     @required Result uninitialized(),
-    @required Result error(String message),
+    @required Result error(String message, String humanMessage),
     @required Result loading(),
-    @required Result loaded(dynamic data, String lastPath, String timestamp),
+    @required
+        Result loaded(Map<String, String> headers, String lastPath,
+            String timestamp, dynamic data),
   }) {
     assert(uninitialized != null);
     assert(error != null);
     assert(loading != null);
     assert(loaded != null);
-    return loaded(data, lastPath, timestamp);
+    return loaded(headers, lastPath, timestamp, data);
   }
 
   @override
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
     Result uninitialized(),
-    Result error(String message),
+    Result error(String message, String humanMessage),
     Result loading(),
-    Result loaded(dynamic data, String lastPath, String timestamp),
+    Result loaded(Map<String, String> headers, String lastPath,
+        String timestamp, dynamic data),
     @required Result orElse(),
   }) {
     assert(orElse != null);
     if (loaded != null) {
-      return loaded(data, lastPath, timestamp);
+      return loaded(headers, lastPath, timestamp, data);
     }
     return orElse();
   }
@@ -453,12 +496,20 @@ class _$LoadedRestState implements LoadedRestState {
 }
 
 abstract class LoadedRestState implements RestState {
-  factory LoadedRestState({dynamic data, String lastPath, String timestamp}) =
-      _$LoadedRestState;
+  factory LoadedRestState(
+      {Map<String, String> headers,
+      String lastPath,
+      String timestamp,
+      dynamic data}) = _$LoadedRestState;
 
-  dynamic get data;
+  Map<String, String> get headers;
   String get lastPath;
   String get timestamp;
+  dynamic get data;
 
-  LoadedRestState copyWith({dynamic data, String lastPath, String timestamp});
+  LoadedRestState copyWith(
+      {Map<String, String> headers,
+      String lastPath,
+      String timestamp,
+      dynamic data});
 }